requirements.txt
Django>=3.0,<4.0
djangorestframework==3.12.4
psycopg2==2.9.1
psycopg2-binary>=2.8
black==21.6b0
flake8==3.9.2
djoser==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-cov==2.12.1
